+++ tests/basics/higher_order/higher_order.cpp
@@ -1,6 +1,4 @@
 #include "higher_order.h"
 
 /// adder n returns a function that adds n to its argument.
-unsigned int HigherOrder::adder(unsigned int _x0, unsigned int _x1) {
-  return (_x0 + _x1);
-}
+uint64_t HigherOrder::adder(uint64_t _x0, uint64_t _x1) { return (_x0 + _x1); }
