+++ tests/basics/tail_rec_zip/tail_rec_zip.t.cpp
@@ -88,10 +88,10 @@ list_to_pairs(const List<Prod<A, B>> &l) {
 int main() {
   // Test 1: Zip empty lists
   {
-    auto empty_a = List<unsigned int>::nil();
-    auto empty_b = List<unsigned int>::nil();
-    auto result = better_zip<unsigned int, unsigned int>(empty_a, empty_b);
-    auto vec = list_to_pairs<unsigned int, unsigned int>(result);
+    auto empty_a = List<uint64_t>::nil();
+    auto empty_b = List<uint64_t>::nil();
+    auto result = better_zip<uint64_t, uint64_t>(empty_a, empty_b);
+    auto vec = list_to_pairs<uint64_t, uint64_t>(result);
     ASSERT(vec.size() == 0);
     std::cout << "Test 1 (zip empty lists): PASSED" << std::endl;
   }
@@ -99,11 +99,11 @@ int main() {
   // Test 2: Zip single elements
   {
     auto a =
-        List<unsigned int>::cons(1, List<unsigned int>::nil());
+        List<uint64_t>::cons(1, List<uint64_t>::nil());
     auto b =
-        List<unsigned int>::cons(2, List<unsigned int>::nil());
-    auto result = better_zip<unsigned int, unsigned int>(a, b);
-    auto vec = list_to_pairs<unsigned int, unsigned int>(result);
+        List<uint64_t>::cons(2, List<uint64_t>::nil());
+    auto result = better_zip<uint64_t, uint64_t>(a, b);
+    auto vec = list_to_pairs<uint64_t, uint64_t>(result);
     ASSERT(vec.size() == 1);
     ASSERT(vec[0].first == 1 && vec[0].second == 2);
     std::cout << "Test 2 (zip single elements): PASSED" << std::endl;
@@ -111,10 +111,10 @@ int main() {
 
   // Test 3: Zip equal length lists
   {
-    auto a = vector_to_list<unsigned int>({1, 2, 3});
-    auto b = vector_to_list<unsigned int>({10, 20, 30});
-    auto result = better_zip<unsigned int, unsigned int>(a, b);
-    auto vec = list_to_pairs<unsigned int, unsigned int>(result);
+    auto a = vector_to_list<uint64_t>({1, 2, 3});
+    auto b = vector_to_list<uint64_t>({10, 20, 30});
+    auto result = better_zip<uint64_t, uint64_t>(a, b);
+    auto vec = list_to_pairs<uint64_t, uint64_t>(result);
     ASSERT(vec.size() == 3);
     ASSERT(vec[0].first == 1 && vec[0].second == 10);
     ASSERT(vec[1].first == 2 && vec[1].second == 20);
@@ -124,10 +124,10 @@ int main() {
 
   // Test 4: Zip with first list shorter
   {
-    auto a = vector_to_list<unsigned int>({1, 2});
-    auto b = vector_to_list<unsigned int>({10, 20, 30, 40});
-    auto result = better_zip<unsigned int, unsigned int>(a, b);
-    auto vec = list_to_pairs<unsigned int, unsigned int>(result);
+    auto a = vector_to_list<uint64_t>({1, 2});
+    auto b = vector_to_list<uint64_t>({10, 20, 30, 40});
+    auto result = better_zip<uint64_t, uint64_t>(a, b);
+    auto vec = list_to_pairs<uint64_t, uint64_t>(result);
     ASSERT(vec.size() == 2);
     ASSERT(vec[0].first == 1 && vec[0].second == 10);
     ASSERT(vec[1].first == 2 && vec[1].second == 20);
@@ -136,10 +136,10 @@ int main() {
 
   // Test 5: Zip with second list shorter
   {
-    auto a = vector_to_list<unsigned int>({1, 2, 3, 4});
-    auto b = vector_to_list<unsigned int>({10, 20});
-    auto result = better_zip<unsigned int, unsigned int>(a, b);
-    auto vec = list_to_pairs<unsigned int, unsigned int>(result);
+    auto a = vector_to_list<uint64_t>({1, 2, 3, 4});
+    auto b = vector_to_list<uint64_t>({10, 20});
+    auto result = better_zip<uint64_t, uint64_t>(a, b);
+    auto vec = list_to_pairs<uint64_t, uint64_t>(result);
     ASSERT(vec.size() == 2);
     ASSERT(vec[0].first == 1 && vec[0].second == 10);
     ASSERT(vec[1].first == 2 && vec[1].second == 20);
@@ -148,10 +148,10 @@ int main() {
 
   // Test 6: Zip with different types
   {
-    auto a = vector_to_list<unsigned int>({1, 2, 3});
+    auto a = vector_to_list<uint64_t>({1, 2, 3});
     auto b = vector_to_list<bool>({true, false, true});
-    auto result = better_zip<unsigned int, bool>(a, b);
-    auto vec = list_to_pairs<unsigned int, bool>(result);
+    auto result = better_zip<uint64_t, bool>(a, b);
+    auto vec = list_to_pairs<uint64_t, bool>(result);
     ASSERT(vec.size() == 3);
     ASSERT(vec[0].first == 1 && vec[0].second == true);
     ASSERT(vec[1].first == 2 && vec[1].second == false);
